+++ examples/channels-example/service/index.js
@@ -1,21 +1,23 @@
 'use strict';
 
-const Hapi = require('hapi');
+const Hapi = require('@hapi/hapi');
 const kafkaRoutes = require('./kafka/kafka-routes');
 
 const port = 8000;
-const server = new Hapi.Server({ port });
+const server = Hapi.server({
+  port,
+});
 
 server.route(kafkaRoutes);
 
 const init = async () => {
-    await server.start();
-    console.log(`Server running at: ${server.info.uri}`);
+  await server.start();
+  console.log(`Server running at: ${server.info.uri}`);
 };
 
 process.on('unhandledRejection', (err) => {
-    console.log(err);
-    process.exit(1);
+  console.log(err);
+  process.exit(1);
 });
 
 init();