+++ examples/channels-example/service/kafka/kafka-ctrl.js
@@ -1,53 +1,33 @@
 'use strict';
 
-const kafka = require('no-kafka');
+const { Kafka } = require('kafkajs');
 
 const KAFKA_HOSTS = process.env.KAFKA_HOSTS || 'localhost:9092';
 const KAFKA_SOURCE_TOPICS = process.env.KAFKA_SOURCE_TOPICS || 'example';
 
-const kafkaProducer = (message) => {
-    const producer = new kafka.Producer({
-        connectionString: KAFKA_HOSTS,
-    });
-
-    const stringMessageValue = JSON.stringify(message);
-
-    return producer.init()
-        .then(() => {
-            const data = {
-                topic: KAFKA_SOURCE_TOPICS,
-                message: {
-                    value: stringMessageValue,
-                },
-            };
-
-            return producer.send(data);
-        })
-        .then((result) => {
-            producer.end();
-
-            const { error } = result[0];
-            if (error) {
-                return error;
-            }
-
-            console.log(`Message successfully produced to Kafka ${JSON.stringify(result)}`);
-            return result;
-        })
-        .catch((e) => {
-            console.log(`Could not produce message to topic ${KAFKA_SOURCE_TOPICS}`);
-            console.log(e);
-            producer.end();
-            return e;
-        });
-};
+const kafka = new Kafka({
+  clientId: 'channel-service',
+  brokers: [KAFKA_HOSTS],
+});
+const producer = kafka.producer();
 
 exports.produce = {
-    handler: (request) => {
-        const msg = request.payload;
+  handler: async (request) => {
+    const msg = request.payload;
 
-        return kafkaProducer(msg)
-            .then(message => ({ status: 'ok', message }))
-            .catch(message => ({ status: 'error', message }));
-    },
+    try {
+      await producer.connect();
+      await producer.send({
+        topic: KAFKA_SOURCE_TOPICS,
+        messages: [{ value: JSON.stringify(msg) }],
+      });
+      await producer.disconnect();
+      console.log('Message successfully produced to Kafka');
+
+      return { status: 'ok' };
+    } catch (error) {
+      console.error('Failed to produce message to Kafka', error);
+      return { status: 'error', message: error };
+    }
+  },
 };